adding second option (supersaturation advection) to adv_hor
